update/ move MyGroups file
