+++ modules/UserCenter/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import ModalWrapper from "@/common/components/ModalWrapper";
 import ProfileSetting from "./ProfileSetting";
-import MyGroupsLeader from "./MyGroupsLeader";
-import MyGroupsMember from "./MyGroupsMember";
+import MyGroupsLeader from "./MyGroups/MyGroupsLeader";
+import MyGroupsMember from "./MyGroups/MyGroupsMember";
 import icons from "@/constants/iconsPackage/userNavIcons";
 import MyFollowing from "./MyFollowing";
 import MyNotification from "./MyNotification";
